pyproject.toml
[project]
name = "gha-hashpinner"
description = """
Update GitHub Actions configurations to use hashpins, instead of mutable pins, in a Dependabot-compatible way
"""
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python",
    "Topic :: Security",
    "Topic :: Utilities",
]
keywords = [
]
dynamic = ["version"]
dependencies = [
    "pygithub>=2.8.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.urls]
Repository = "https://github.com/mfisher87/gha-hashpinner.git"
Homepage = "https://github.com/mfisher87/gha-hashpinner"
Issues = "https://github.com/mfisher87/gha-hashpinner/issues"

[project.scripts]
gha-hashpinner = "gha_hashpinner.cli:cli"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "types-pyyaml>=6.0.12.20250915",
]
test = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]


[build-system]
requires = ["hatchling>=1.5.0", "hatch-vcs", "jupyterlab>=4.0.0,<5"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.sdist]
exclude = [".github"]

[tool.hatch.build.hooks.version]
path = "src/gha_hashpinner/_version.py"


[tool.uv]
# "Dependency cooldown" to help prevent supply chain attacks
#    See: https://blog.yossarian.net/2025/11/21/We-should-all-be-using-dependency-cooldowns
exclude-newer = "7 days"


[tool.mypy]
packages = ["gha_hashpinner", "tests"]
python_version = "3.12"
strict = true


[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN002",  # Require type annotation for *args
    "ANN003",  # Require type annotation for **kwargs
    "D203",    # Require blank line between class signature and docstring
    "D213",    # Require multi-line docstring summary on 2nd line
    "COM812",  # Require trailing commas (the Ruff formatter does this)
    "EM101",   # Forbid literal strings in exception messages
    "EM102",   # Forbid f-strings in exception messages
    "FIX",     # FIXME, TODO, etc. comments
    "G004",    # Forbid f-strings in log messages
    "INP",     # Forbid implicit namespace packages
    "TD",      # Require standards for TODO comments
    "TRY003",  # Forbid "long" exception messages
]

[tool.ruff.lint.per-file-ignores]
"{**/tests/**/*.py}" = [
    "ARG001",   # Forbid unused function arguments (e.g. test fixtures)
    "D101",     # Require docstrings for every method
    "E501",     # Limit line length
    "PLR2004",  # Forbid magic value in comparisons
    "RUF015",   # Prefer next() for single-item slice
    "S101",     # Forbid assertion statements
    "S106",     # Forbid hardcoded passwords (in tests, we use dummy tokens)
    "SLF001",   # Forbid private member access
]
